final soa

+++ device/common/include/traccc/clusterization/device/aggregate_cluster.hpp
@@ -30,10 +30,10 @@ namespace traccc::device {
 /// @param[out] out     cluster to fill
 TRACCC_HOST_DEVICE
 inline void aggregate_cluster(
-    vecmem::data::vector_view<unsigned short> ch0 ,
-     vecmem::data::vector_view<unsigned short> ch1, 
+    vecmem::data::vector_view<unsigned int> ch0 ,
+     vecmem::data::vector_view<unsigned int> ch1, 
      vecmem::data::vector_view<scalar> activation ,
-     vecmem::data::vector_view<unsigned short> module_link ,
+     vecmem::data::vector_view<unsigned int> module_link ,
     const cell_module_collection_types::const_device& modules,
     const vecmem::data::vector_view<unsigned short> f_view,
     const unsigned int start, const unsigned int end, const unsigned short cid,
